添加Sphere，Disk，Triangle和SSAA抗锯齿

+++ src/Camera.cpp
@@ -31,7 +31,7 @@ void Camera::Initialize(const Vector3f& p, const Vector3f& target, const Vector3
 	mInverseCombinedMatrix = inverseCombinedMatrix;
 }
 
-Ray Camera::GetRay(int x, int y) const
+Ray Camera::GetRay(float x, float y) const
 {
 	Ray ray;
 	ray.o = mPosition;
